requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2

